+++ lib/spreedly/environment.rb
@@ -326,6 +326,7 @@ module Spreedly
     def add_extra_options_for_basic_ops(doc, options)
       add_gateway_specific_fields(doc, options)
       add_shipping_address_override(doc, options)
+      add_billing_address_override(doc, options)
       add_to_doc(doc, options, :order_id, :description, :ip, :email, :merchant_name_descriptor,
                                :merchant_location_descriptor, :redirect_url, :callback_url,
                                :continue_caching, :attempt_3dsecure, :browser_info, :three_ds_version, :channel)
@@ -340,6 +341,15 @@ module Spreedly
       return unless options[:shipping_address].kind_of?(Hash)
       doc.send(:shipping_address) do
         options[:shipping_address].each do |k, v|
+          doc.send(k, v)
+        end
+      end
+    end
+
+    def add_billing_address_override(doc, options)
+      return unless options[:billing_address].kind_of?(Hash)
+      doc.send(:billing_address) do
+        options[:billing_address].each do |k, v|
           doc.send(k, v)
         end
       end